Add customize-cra react-app-rewired to add images from public folder - Refactor some code
