+++ frontend/config-overrides.js
@@ -1,0 +1,3 @@
+const { removeModuleScopePlugin } = require("customize-cra");
+
+module.exports = removeModuleScopePlugin();
