+++ frontend/public/images/airpods.jpg
