+++ frontend/public/images/sofa.png
