+++ frontend/public/images/square_active.png
