+++ frontend/src/components/Homepage/LeatestProducts/Items/IconsCarts.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import { CartIconPosition, IconInteractive } from "../../style/Items.style";
-import cartIcon from "../../../../assets/images/icons/cart-icon.svg";
-import heartIcon from "../../../../assets/images/icons/heart-icon.svg";
-import increaseImage from "../../../../assets/images/icons/increase-image.svg";
+import cartIcon from "../../../../../public/images/icons/cart-icon.svg";
+import heartIcon from "../../../../../public/images/icons/heart-icon.svg";
+import increaseImage from "../../../../../public/images/icons/increase-image.svg";
 
 interface IIconsCartsProps {
   handleAddToCart: () => void;
