+++ frontend/src/store/actions/logoutTimer.tsx
@@ -7,3 +7,5 @@ export const logoutTimer = (time: number) => {
     }, time);
   };
 };
+
+//Da fare su server
